Its workinggg
The animation is working and its looking pretty good, the Animatedsprite2d while yes painful to setup works amazingly in switching between my needed animations for the sprites.
Its also nice that for the transition from black to colour(talking about going from the sihlouhete to full colour) I just put a rectangle to hide the transition and it works amazingly.